Quick Sort Added
This commit introduces the implementation of the Quick Sort algorithm, using the last element of the array as the pivot. Quick Sort is a divide-and-conquer algorithm that sorts an array by partitioning it into two sub-arrays based on the pivot value, then recursively sorting the sub-arrays. The selection of the last element as the pivot is a common strategy that helps balance the partitions in average cases. This implementation improves the sorting functionality in the project, offering an efficient O(n log n) average time complexity.